chore: segregate basic access to fields and more advanced logic
Move basic access to ENR fields (with conversion) to `RawEnr`

`ENR` inherits from `RawEnr` and add a layer of logic.
